refactor: rename router
to match with REST
